EOS-17458 : CortxV2-SSPL-Enhance-System-Manifest
