+++ unittests/conftest.py
@@ -6,4 +6,3 @@ BASE_DIR = os.path.dirname(os.path.dirname(os.path.abspath(__file__)))
 
 sys.path.append(os.path.join(BASE_DIR, 'low-level'))
 sys.path.append(os.path.join(BASE_DIR, 'low-level/files/opt/seagate/sspl/setup/resource_map/lr2'))
-sys.path.append(os.path.join(BASE_DIR, 'low-level/files/opt/seagate/sspl/setup/cortx_manifest/lr2'))
