--- unittests/files/opt/seagate/sspl/setup/cortx_manifest/lr2/encl_api_response.py
@@ -1,1136 +1,0 @@
-#!/usr/bin/python3.6
-
-# Copyright (c) 2021 Seagate Technology LLC and/or its Affiliates
-# This program is free software: you can redistribute it and/or modify
-# it under the terms of the GNU Affero General Public License as published
-# by the Free Software Foundation, either version 3 of the License, or
-# (at your option) any later version.
-# This program is distributed in the hope that it will be useful,
-# but WITHOUT ANY WARRANTY; without even the implied warranty of
-# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
-# GNU Affero General Public License for more details.
-# You should have received a copy of the GNU Affero General Public License
-# along with this program. If not, see <https://www.gnu.org/licenses/>.
-# For any questions about this software or licensing,
-# please email opensource@seagate.com or cortx-questions@seagate.com
-
-
-ENCLOSURE_RESPONSE = """
-    {
-    "status_code": 200,
-    "api-response": {
-            "enclosures": [
-            {
-                "object-name": "enclosures",
-                "durable-id": "enclosure_0",
-                "enclosure-id": 0,
-                "enclosure-wwn": "500C0FF03B98E23C",
-                "name": "",
-                "type": "Titan",
-                "type-numeric": 12,
-                "iom-type": "Xyratex5U84Rbod",
-                "iom-type-numeric": 4,
-                "platform-type": "Gallium3 NX",
-                "platform-type-numeric": 8,
-                "board-model": "Gallium Raidhead-12G",
-                "board-model-numeric": 11,
-                "location": "",
-                "rack-number": 0,
-                "rack-position": 0,
-                "number-of-coolings-elements": 10,
-                "number-of-disks": 28,
-                "number-of-power-supplies": 2,
-                "status": "Unknown",
-                "status-numeric": 6,
-                "extended-status": "00000000",
-                "midplane-serial-number": "DHSIHOU-18103B98E2",
-                "vendor": "Seagate",
-                "model": "SP-3584-GAL3-NX",
-                "fru-shortname": "",
-                "fru-location": "MID-PLANE SLOT",
-                "part-number": "FRUKA62-01",
-                "mfg-date": "2018-01-08 10:31:00",
-                "mfg-date-numeric": 1515407460,
-                "mfg-location": "",
-                "description": "N/A",
-                "revision": "C",
-                "dash-level": "",
-                "emp-a-rev": "513E",
-                "emp-b-rev": "513E",
-                "rows": 3,
-                "columns": 14,
-                "slots": 84,
-                "locator-led": "Off",
-                "locator-led-numeric": 0,
-                "drive-orientation": "horizontal",
-                "drive-orientation-numeric": 1,
-                "enclosure-arrangement": "vertical",
-                "enclosure-arrangement-numeric": 0,
-                "emp-a-busid": "00",
-                "emp-a-targetid": "127",
-                "emp-b-busid": "01",
-                "emp-b-targetid": "127",
-                "emp-a": "",
-                "emp-a-ch-id-rev": "00:127 513E",
-                "emp-b": "",
-                "emp-b-ch-id-rev": "01:127 513E",
-                "midplane-type": "5U84-12G",
-                "midplane-type-numeric": 71,
-                "midplane-rev": 0,
-                "enclosure-power": "666.78",
-                "pcie2-capable": "False",
-                "pcie2-capable-numeric": 0,
-                "health": "Degraded",
-                "health-numeric": 1,
-                "health-reason": "Health is not applicable in this situation.",
-                "health-recommendation": "- No action is required.",
-                "drawers": [
-                {
-                    "object-name": "drawers",
-                    "durable-id": "drawer_0.0",
-                    "drawer-id": 0,
-                    "drawer-wwn": "0000000000000000",
-                    "part-number": "N/A",
-                    "name": "Drawer 0",
-                    "position": "Top",
-                    "position-numeric": 2,
-                    "rows": 3,
-                    "columns": 14,
-                    "slots": 42,
-                    "number-of-disks": 14,
-                    "emp-a-busid": "N/A",
-                    "emp-a-targetid": "N/A",
-                    "emp-a-rev": "N/A",
-                    "emp-b-busid": "N/A",
-                    "emp-b-targetid": "N/A",
-                    "emp-b-rev": "N/A",
-                    "emp-a": "A",
-                    "emp-a-ch-id-rev": "N/A",
-                    "emp-b": "B",
-                    "emp-b-ch-id-rev": "N/A",
-                    "locator-led": "Off",
-                    "locator-led-numeric": 0,
-                    "status": "OK",
-                    "status-numeric": 1,
-                    "extended-status": "00000000",
-                    "health": "OK",
-                    "health-numeric": 0,
-                    "health-reason": "",
-                    "health-recommendation": "",
-                    "sideplanes": [
-                    {
-                        "object-name": "sideplanes",
-                        "durable-id": "sideplane_0.D0.B",
-                        "enclosure-id": 0,
-                        "drawer-id": 0,
-                        "dom-id": 1,
-                        "path-id": "B",
-                        "path-id-numeric": 0,
-                        "name": "Left Sideplane",
-                        "location": "enclosure 0, drawer 0",
-                        "position": "Right",
-                        "position-numeric": 1,
-                        "status": "OK",
-                        "status-numeric": 1,
-                        "extended-status": "00000000",
-                        "health": "OK",
-                        "health-numeric": 0,
-                        "health-reason": "",
-                        "health-recommendation": "",
-                        "expanders": [
-                        {
-                            "object-name": "expanders",
-                            "durable-id": "expander_0.D0.B0",
-                            "enclosure-id": 0,
-                            "drawer-id": 0,
-                            "dom-id": 0,
-                            "path-id": "B",
-                            "path-id-numeric": 0,
-                            "name": "Sideplane 24-port Expander 0",
-                            "location": "Enclosure 0, Drawer 0, Left Sideplane",
-                            "status": "OK",
-                            "status-numeric": 1,
-                            "extended-status": "00000000",
-                            "fw-revision": "513e",
-                            "health": "OK",
-                            "health-numeric": 0,
-                            "health-reason": "",
-                            "health-recommendation": ""
-                        }
-                        ]
-                    }
-                    ]
-                }
-                ]
-            }
-            ],
-            "status": [
-            {
-                "object-name": "status",
-                "response-type": "Success",
-                "response-type-numeric": 0,
-                "response": "Command completed successfully. (2021-06-21 10:52:14)",
-                "return-code": 0,
-                "component-id": "",
-                "time-stamp": "2021-06-21 10:52:14",
-                "time-stamp-numeric": 1624272734
-            }
-        ]
-    }}
-    """
-
-CONTROLLER_RESPONSE = """
-    {
-    "status_code": 200,
-    "api-response": {
-        "controllers":[
-        {
-        "object-name":"controllers",
-            "durable-id":"controller_a",
-            "controller-id":"A",
-            "controller-id-numeric":1,
-            "serial-number":"DHSIFTJ-18253C638B",
-            "hardware-version":"1.1",
-            "cpld-version":"3.9",
-            "mac-address":"00:C0:FF:3C:63:8B",
-            "node-wwn":"208000C0FFA7117A",
-            "active-version":4,
-            "ip-address":"10.230.163.30",
-            "ip-subnet-mask":"255.255.248.0",
-            "ip-gateway":"10.230.160.1",
-            "disks":84,
-            "number-of-storage-pools":0,
-            "virtual-disks":2,
-            "cache-memory-size":4096,
-            "system-memory-size":8192,
-            "host-ports":2,
-            "drive-channels":2,
-            "drive-bus-type":"SAS",
-            "drive-bus-type-numeric":8,
-            "status":"Operational",
-            "status-numeric":0,
-            "failed-over":"No",
-            "failed-over-numeric":0,
-            "fail-over-reason":"Not applicable",
-            "fail-over-reason-numeric":0,
-            "sc-fw":"GTS265R18-01",
-            "vendor":"",
-            "model":"3865",
-            "platform-type":"Gallium3 EX",
-            "platform-type-numeric":7,
-            "multicore":"Enabled",
-            "multicore-numeric":1,
-            "sc-cpu-type":"Broadwell",
-            "sc-cpu-speed":1200,
-            "internal-serial-number":"00C0FF3C638B",
-            "cache-lock":"Yes",
-            "cache-lock-numeric":1,
-            "write-policy":"write-back",
-            "write-policy-numeric":0,
-            "description":"N/A",
-            "part-number":"81-00000117-00-15",
-            "revision":"A",
-            "dash-level":"",
-            "fru-shortname":"",
-            "mfg-date":"2007-01-03 16:24:00",
-            "mfg-date-numeric":1167841440,
-            "mfg-location":"",
-            "mfg-vendor-id":"",
-            "locator-led":"Off",
-            "locator-led-numeric":0,
-            "ssd-alt-path-io-count":2,
-            "health":"OK",
-            "health-numeric":0,
-            "health-reason":"",
-            "health-recommendation":"",
-            "position":"Left",
-            "position-numeric":0,
-            "rotation":"0 Degrees",
-            "rotation-numeric":0,
-            "phy-isolation":"Enabled",
-            "phy-isolation-numeric":0,
-            "redundancy-mode":"Active-Active ULP",
-            "redundancy-mode-numeric":8,
-            "redundancy-status":"Redundant",
-            "redundancy-status-numeric":2,
-            "network-parameters":[
-            {
-            "object-name":"controller-a",
-            "durable-id":"mgmtport_a",
-            "active-version":4,
-            "ip-address":"10.230.163.30",
-            "gateway":"10.230.160.1",
-            "subnet-mask":"255.255.248.0",
-            "mac-address":"00:c0:ff:3c:63:8b",
-            "addressing-mode":"DHCP",
-            "addressing-mode-numeric":2,
-            "link-speed":"1000mbps",
-            "link-speed-numeric":2,
-            "duplex-mode":"full",
-            "duplex-mode-numeric":0,
-            "auto-negotiation":"Enabled",
-            "auto-negotiation-numeric":1,
-            "health":"OK",
-            "health-numeric":0,
-            "health-reason":"",
-            "health-recommendation":"",
-            "ping-broadcast":"Disabled",
-            "ping-broadcast-numeric":0
-            }
-        ],
-            "port":[
-            {
-            "object-name":"ports",
-            "durable-id":"hostport_A0",
-            "controller":"A",
-            "controller-numeric":1,
-            "port":"A0",
-            "port-type":"FC",
-            "port-type-numeric":6,
-            "media":"FC(P)",
-            "target-id":"207000c0ffa7117a",
-            "status":"Up",
-            "status-numeric":0,
-            "actual-speed":"8Gb",
-            "actual-speed-numeric":7,
-            "configured-speed":"Auto",
-            "configured-speed-numeric":3,
-            "fan-out":1,
-            "health":"OK",
-            "health-numeric":0,
-            "health-reason":"",
-            "health-recommendation":"",
-            "fc-port":[
-            {
-                "object-name":"port-details",
-                "configured-topology":"PTP",
-                "configured-topology-numeric":1,
-                "primary-loop-id":"N/A",
-                "sfp-status":"OK",
-                "sfp-status-numeric":3,
-                "sfp-present":"Present",
-                "sfp-present-numeric":1,
-                "sfp-vendor":"DHS",
-                "sfp-part-number":"FTLF8529P3BCV-1D",
-                "sfp-revision":"A",
-                "sfp-supported-speeds":"4G,8G,16G",
-                "sfp-supported-speeds-numeric":112
-            }
-            ]
-            },
-            {
-            "object-name":"ports",
-            "durable-id":"hostport_A1",
-            "controller":"A",
-            "controller-numeric":1,
-            "port":"A1",
-            "port-type":"FC",
-            "port-type-numeric":6,
-            "media":"FC(P)",
-            "target-id":"217000c0ffa7117a",
-            "status":"Up",
-            "status-numeric":0,
-            "actual-speed":"8Gb",
-            "actual-speed-numeric":7,
-            "configured-speed":"Auto",
-            "configured-speed-numeric":3,
-            "fan-out":1,
-            "health":"OK",
-            "health-numeric":0,
-            "health-reason":"",
-            "health-recommendation":"",
-            "fc-port":[
-            {
-                "object-name":"port-details",
-                "configured-topology":"PTP",
-                "configured-topology-numeric":1,
-                "primary-loop-id":"N/A",
-                "sfp-status":"OK",
-                "sfp-status-numeric":3,
-                "sfp-present":"Present",
-                "sfp-present-numeric":1,
-                "sfp-vendor":"DHS",
-                "sfp-part-number":"FTLF8529P3BCV-1D",
-                "sfp-revision":"A",
-                "sfp-supported-speeds":"4G,8G,16G",
-                "sfp-supported-speeds-numeric":112
-            }
-            ]
-            }
-        ],
-            "expander-ports":[
-            {
-            "object-name":"expander-port",
-            "durable-id":"expport_universal_0.A0",
-            "enclosure-id":0,
-            "controller":"A",
-            "controller-numeric":1,
-            "sas-port-type":"Expansion Port Universal",
-            "sas-port-type-numeric":5,
-            "sas-port-index":0,
-            "name":"Expansion Port",
-            "status":"Disconnected",
-            "status-numeric":6,
-            "health":"N/A",
-            "health-numeric":4,
-            "health-reason":"No drive enclosure is connected to this expansion port. This is normal if this is the last (or only) enclosure in the cabling sequence of the system.",
-            "health-recommendation":"- No action is required."
-            },
-            {
-            "object-name":"expander-port",
-            "durable-id":"drawer_egress_0.D0.A0",
-            "enclosure-id":0,
-            "controller":"A",
-            "controller-numeric":1,
-            "sas-port-type":"Drawer Port Egress",
-            "sas-port-type-numeric":1,
-            "sas-port-index":0,
-            "name":"Drawer0-egress A0",
-            "status":"Up",
-            "status-numeric":0,
-            "health":"OK",
-            "health-numeric":0,
-            "health-reason":"",
-            "health-recommendation":""
-            },
-            {
-            "object-name":"expander-port",
-            "durable-id":"drawer_egress_0.D0.A1",
-            "enclosure-id":0,
-            "controller":"A",
-            "controller-numeric":1,
-            "sas-port-type":"Drawer Port Egress",
-            "sas-port-type-numeric":1,
-            "sas-port-index":1,
-            "name":"Drawer0-egress A1",
-            "status":"Up",
-            "status-numeric":0,
-            "health":"OK",
-            "health-numeric":0,
-            "health-reason":"",
-            "health-recommendation":""
-            },
-            {
-            "object-name":"expander-port",
-            "durable-id":"drawer_egress_0.D0.A2",
-            "enclosure-id":0,
-            "controller":"A",
-            "controller-numeric":1,
-            "sas-port-type":"Drawer Port Egress",
-            "sas-port-type-numeric":1,
-            "sas-port-index":2,
-            "name":"Drawer0-egress A2",
-            "status":"Up",
-            "status-numeric":0,
-            "health":"OK",
-            "health-numeric":0,
-            "health-reason":"",
-            "health-recommendation":""
-            },
-            {
-            "object-name":"expander-port",
-            "durable-id":"drawer_egress_0.D1.A0",
-            "enclosure-id":0,
-            "controller":"A",
-            "controller-numeric":1,
-            "sas-port-type":"Drawer Port Egress",
-            "sas-port-type-numeric":1,
-            "sas-port-index":0,
-            "name":"Drawer1-egress A0",
-            "status":"Up",
-            "status-numeric":0,
-            "health":"OK",
-            "health-numeric":0,
-            "health-reason":"",
-            "health-recommendation":""
-            },
-            {
-            "object-name":"expander-port",
-            "durable-id":"drawer_egress_0.D1.A1",
-            "enclosure-id":0,
-            "controller":"A",
-            "controller-numeric":1,
-            "sas-port-type":"Drawer Port Egress",
-            "sas-port-type-numeric":1,
-            "sas-port-index":1,
-            "name":"Drawer1-egress A1",
-            "status":"Up",
-            "status-numeric":0,
-            "health":"OK",
-            "health-numeric":0,
-            "health-reason":"",
-            "health-recommendation":""
-            },
-            {
-            "object-name":"expander-port",
-            "durable-id":"drawer_egress_0.D1.A2",
-            "enclosure-id":0,
-            "controller":"A",
-            "controller-numeric":1,
-            "sas-port-type":"Drawer Port Egress",
-            "sas-port-type-numeric":1,
-            "sas-port-index":2,
-            "name":"Drawer1-egress A2",
-            "status":"Up",
-            "status-numeric":0,
-            "health":"OK",
-            "health-numeric":0,
-            "health-reason":"",
-            "health-recommendation":""
-            }
-        ],
-            "compact-flash":[
-            {
-            "object-name":"controller-a-compact-flash",
-            "durable-id":"Ctlr A CF",
-            "controller-id":"A",
-            "controller-id-numeric":1,
-            "name":"Controller A CompactFlash",
-            "status":"Installed",
-            "status-numeric":1,
-            "cache-flush":"Enabled",
-            "cache-flush-numeric":1,
-            "health":"OK",
-            "health-numeric":0,
-            "health-reason":"",
-            "health-recommendation":""
-            }
-        ],
-            "expanders":[
-            {
-            "object-name":"expanders",
-            "durable-id":"expander_0.A",
-            "enclosure-id":0,
-            "drawer-id":255,
-            "dom-id":0,
-            "path-id":"A",
-            "path-id-numeric":1,
-            "name":"Root Expander-A",
-            "location":"Enclosure 0, Iom A",
-            "status":"OK",
-            "status-numeric":1,
-            "extended-status":"00000000",
-            "fw-revision":"5140",
-            "health":"OK",
-            "health-numeric":0,
-            "health-reason":"",
-            "health-recommendation":""
-            }
-        ]
-        }
-        ],
-        "status":[
-        {
-            "object-name":"status",
-            "response-type":"Success",
-            "response-type-numeric":0,
-            "response":"Command completed successfully. (2019-07-29 11:56:34)",
-            "return-code":0,
-            "component-id":"",
-            "time-stamp":"2019-07-29 11:56:34",
-            "time-stamp-numeric":1564401394
-        }
-        ]
-    }}
-    """
-
-DRIVE_RESPONSE ="""
-    {
-        "status_code": 200,
-        "api-response": {
-            "drives": [
-            {
-                "object-name":"drive",
-                "durable-id":"disk_00.00",
-                "enclosure-id":0,
-                "drawer-id":0,
-                "slot":0,
-                "location":"0.0",
-                "port":0,
-                "scsi-id":0,
-                "blocksize":512,
-                "blocks":3907029168,
-                "serial-number":"Z4H099ZE0000R6375N70",
-                "vendor":"SEAGATE",
-                "model":"ST2000NM0034",
-                "revision":"E0B1",
-                "secondary-channel":0,
-                "container-index":0,
-                "member-index":0,
-                "description":"SAS MDL",
-                "description-numeric":11,
-                "architecture":"HDD",
-                "architecture-numeric":1,
-                "interface":"SAS",
-                "interface-numeric":0,
-                "single-ported":"Disabled",
-                "single-ported-numeric":0,
-                "type":"SAS MDL",
-                "type-numeric":11,
-                "usage":"LINEAR POOL",
-                "usage-numeric":1,
-                "job-running":"",
-                "job-running-numeric":0,
-                "state":"LINEAR POOL",
-                "current-job-completion":"",
-                "blink":0,
-                "locator-led":"Off",
-                "locator-led-numeric":0,
-                "speed":0,
-                "smart":"Enabled",
-                "smart-numeric":1,
-                "dual-port":1,
-                "error":0,
-                "fc-p1-channel":0,
-                "fc-p1-device-id":0,
-                "fc-p1-node-wwn":"5000C500856867D7",
-                "fc-p1-port-wwn":"0000000000000000",
-                "fc-p1-unit-number":0,
-                "fc-p2-channel":0,
-                "fc-p2-device-id":0,
-                "fc-p2-node-wwn":"",
-                "fc-p2-port-wwn":"",
-                "fc-p2-unit-number":0,
-                "drive-down-code":0,
-                "owner":"A",
-                "owner-numeric":1,
-                "index":0,
-                "rpm":7,
-                "size":"2000.3GB",
-                "size-numeric":3907029168,
-                "sector-format":"512e",
-                "sector-format-numeric":1,
-                "transfer-rate":"6.0",
-                "transfer-rate-numeric":3,
-                "attributes":" ",
-                "attributes-numeric":2,
-                "enclosure-wwn":"500C0FF0A7117ABF",
-                "status":"Up",
-                "recon-state":"N/A",
-                "recon-state-numeric":0,
-                "copyback-state":"N/A",
-                "copyback-state-numeric":0,
-                "virtual-disk-serial":"00c0ff3c638b0000a803955c00000000",
-                "disk-group":"poola",
-                "storage-pool-name":"poola",
-                "storage-tier":"Archive",
-                "storage-tier-numeric":4,
-                "ssd-life-left":"N/A",
-                "ssd-life-left-numeric":255,
-                "led-status":" Online",
-                "led-status-numeric":0,
-                "disk-dsd-count":1,
-                "spun-down":0,
-                "number-of-ios":0,
-                "total-data-transferred":"0B",
-                "total-data-transferred-numeric":0,
-                "avg-rsp-time":0,
-                "fde-state":"Not FDE Capable",
-                "fde-state-numeric":1,
-                "lock-key-id":"00000000",
-                "import-lock-key-id":"00000000",
-                "fde-config-time":"N/A",
-                "fde-config-time-numeric":0,
-                "temperature":"20 C",
-                "temperature-numeric":20,
-                "temperature-status":"OK",
-                "temperature-status-numeric":1,
-                "pi-formatted":"No",
-                "pi-formatted-numeric":0,
-                "power-on-hours":2847,
-                "extended-status":0,
-                "health":"OK",
-                "health-numeric":0,
-                "health-reason":"",
-                "health-recommendation":""
-            }
-            ],
-        "status": [
-            {
-                "object-name": "status",
-                "response-type": "Info",
-                "response-type-numeric": 2,
-                "response": "* Rates may vary. This is normal behavior. (2019-06-27 04:24:10)",
-                "return-code": 0,
-                "component-id": "",
-                "time-stamp": "2019-06-27 04:24:10",
-                "time-stamp-numeric": 1561609450
-            },
-            {
-                "object-name": "status",
-                "response-type": "Success",
-                "response-type-numeric": 0,
-                "response": "Command completed successfully. (2019-06-27 04:24:10)",
-                "return-code": 0,
-                "component-id": "",
-                "time-stamp": "2019-06-27 04:24:10",
-                "time-stamp-numeric": 1561609450
-            }
-        ]
-    }}
-    """
-
-FAN_MODULE_RESPONSE = """
-    {
-	"status_code": 200,
-	"api-response": {
-		"fan-modules": [
-			{
-				"object-name": "fan-modules",
-				"meta": "/meta/fan-modules",
-				"durable-id": "fan_module_0.0",
-				"enclosure-id": 0,
-				"dom-id": 0,
-				"name": "Fan Module 0",
-				"location": "Enclosure 0 - Left",
-				"status": "OK",
-				"status-numeric": 1,
-				"extended-status": "00000000",
-				"position": "Indexed",
-				"position-numeric": 5,
-				"health": "OK",
-				"health-numeric": 0,
-				"health-reason": "",
-				"health-recommendation": "",
-				"fan": [
-					{
-						"object-name": "fan-details",
-						"meta": "/meta/fan",
-						"durable-id": "fan_0.fm0.0",
-						"name": "Fan 0",
-						"location": "Enclosure 0, Fan Module 0",
-						"status-ses": "OK",
-						"status-ses-numeric": 1,
-						"extended-status": "00000010",
-						"status": "Up",
-						"status-numeric": 0,
-						"speed": 13800,
-						"position": "N/A",
-						"position-numeric": 6,
-						"serial-number": "N/A",
-						"part-number": "N/A",
-						"fw-revision": "",
-						"hw-revision": "",
-						"locator-led": "Off",
-						"locator-led-numeric": 0,
-						"health": "OK",
-						"health-numeric": 0,
-						"health-reason": "",
-						"health-recommendation": ""
-					},
-					{
-						"object-name": "fan-details",
-						"meta": "/meta/fan",
-						"durable-id": "fan_0.fm0.1",
-						"name": "Fan 1",
-						"location": "Enclosure 0, Fan Module 0",
-						"status-ses": "OK",
-						"status-ses-numeric": 1,
-						"extended-status": "00000010",
-						"status": "Up",
-						"status-numeric": 0,
-						"speed": 13860,
-						"position": "N/A",
-						"position-numeric": 6,
-						"serial-number": "N/A",
-						"part-number": "N/A",
-						"fw-revision": "",
-						"hw-revision": "",
-						"locator-led": "Off",
-						"locator-led-numeric": 0,
-						"health": "OK",
-						"health-numeric": 0,
-						"health-reason": "",
-						"health-recommendation": ""
-					}
-				]
-			},
-			{
-				"object-name": "fan-modules",
-				"meta": "/meta/fan-modules",
-				"durable-id": "fan_module_0.1",
-				"enclosure-id": 0,
-				"dom-id": 1,
-				"name": "Fan Module 1",
-				"location": "Enclosure 0 - Left-Middle",
-				"status": "OK",
-				"status-numeric": 1,
-				"extended-status": "00000000",
-				"position": "Indexed",
-				"position-numeric": 5,
-				"health": "OK",
-				"health-numeric": 0,
-				"health-reason": "",
-				"health-recommendation": "",
-				"fan": [
-					{
-						"object-name": "fan-details",
-						"meta": "/meta/fan",
-						"durable-id": "fan_0.fm1.0",
-						"name": "Fan 0",
-						"location": "Enclosure 0, Fan Module 1",
-						"status-ses": "OK",
-						"status-ses-numeric": 1,
-						"extended-status": "00000010",
-						"status": "Up",
-						"status-numeric": 0,
-						"speed": 13860,
-						"position": "N/A",
-						"position-numeric": 6,
-						"serial-number": "N/A",
-						"part-number": "N/A",
-						"fw-revision": "",
-						"hw-revision": "",
-						"locator-led": "Off",
-						"locator-led-numeric": 0,
-						"health": "OK",
-						"health-numeric": 0,
-						"health-reason": "",
-						"health-recommendation": ""
-					},
-					{
-						"object-name": "fan-details",
-						"meta": "/meta/fan",
-						"durable-id": "fan_0.fm1.1",
-						"name": "Fan 1",
-						"location": "Enclosure 0, Fan Module 1",
-						"status-ses": "OK",
-						"status-ses-numeric": 1,
-						"extended-status": "00000010",
-						"status": "Up",
-						"status-numeric": 0,
-						"speed": 13500,
-						"position": "N/A",
-						"position-numeric": 6,
-						"serial-number": "N/A",
-						"part-number": "N/A",
-						"fw-revision": "",
-						"hw-revision": "",
-						"locator-led": "Off",
-						"locator-led-numeric": 0,
-						"health": "OK",
-						"health-numeric": 0,
-						"health-reason": "",
-						"health-recommendation": ""
-					}
-				]
-			},
-			{
-				"object-name": "fan-modules",
-				"meta": "/meta/fan-modules",
-				"durable-id": "fan_module_0.2",
-				"enclosure-id": 0,
-				"dom-id": 2,
-				"name": "Fan Module 2",
-				"location": "Enclosure 0 - Middle",
-				"status": "OK",
-				"status-numeric": 1,
-				"extended-status": "00000000",
-				"position": "Indexed",
-				"position-numeric": 5,
-				"health": "OK",
-				"health-numeric": 0,
-				"health-reason": "",
-				"health-recommendation": "",
-				"fan": [
-					{
-						"object-name": "fan-details",
-						"meta": "/meta/fan",
-						"durable-id": "fan_0.fm2.0",
-						"name": "Fan 0",
-						"location": "Enclosure 0, Fan Module 2",
-						"status-ses": "OK",
-						"status-ses-numeric": 1,
-						"extended-status": "00000010",
-						"status": "Up",
-						"status-numeric": 0,
-						"speed": 13860,
-						"position": "N/A",
-						"position-numeric": 6,
-						"serial-number": "N/A",
-						"part-number": "N/A",
-						"fw-revision": "",
-						"hw-revision": "",
-						"locator-led": "Off",
-						"locator-led-numeric": 0,
-						"health": "OK",
-						"health-numeric": 0,
-						"health-reason": "",
-						"health-recommendation": ""
-					},
-					{
-						"object-name": "fan-details",
-						"meta": "/meta/fan",
-						"durable-id": "fan_0.fm2.1",
-						"name": "Fan 1",
-						"location": "Enclosure 0, Fan Module 2",
-						"status-ses": "OK",
-						"status-ses-numeric": 1,
-						"extended-status": "00000010",
-						"status": "Up",
-						"status-numeric": 0,
-						"speed": 13860,
-						"position": "N/A",
-						"position-numeric": 6,
-						"serial-number": "N/A",
-						"part-number": "N/A",
-						"fw-revision": "",
-						"hw-revision": "",
-						"locator-led": "Off",
-						"locator-led-numeric": 0,
-						"health": "OK",
-						"health-numeric": 0,
-						"health-reason": "",
-						"health-recommendation": ""
-					}
-				]
-			},
-			{
-				"object-name": "fan-modules",
-				"meta": "/meta/fan-modules",
-				"durable-id": "fan_module_0.3",
-				"enclosure-id": 0,
-				"dom-id": 3,
-				"name": "Fan Module 3",
-				"location": "Enclosure 0 - Right-Middle",
-				"status": "OK",
-				"status-numeric": 1,
-				"extended-status": "00000000",
-				"position": "Indexed",
-				"position-numeric": 5,
-				"health": "OK",
-				"health-numeric": 0,
-				"health-reason": "",
-				"health-recommendation": "",
-				"fan": [
-					{
-						"object-name": "fan-details",
-						"meta": "/meta/fan",
-						"durable-id": "fan_0.fm3.0",
-						"name": "Fan 0",
-						"location": "Enclosure 0, Fan Module 3",
-						"status-ses": "OK",
-						"status-ses-numeric": 1,
-						"extended-status": "00000010",
-						"status": "Up",
-						"status-numeric": 0,
-						"speed": 13860,
-						"position": "N/A",
-						"position-numeric": 6,
-						"serial-number": "N/A",
-						"part-number": "N/A",
-						"fw-revision": "",
-						"hw-revision": "",
-						"locator-led": "Off",
-						"locator-led-numeric": 0,
-						"health": "OK",
-						"health-numeric": 0,
-						"health-reason": "",
-						"health-recommendation": ""
-					},
-					{
-						"object-name": "fan-details",
-						"meta": "/meta/fan",
-						"durable-id": "fan_0.fm3.1",
-						"name": "Fan 1",
-						"location": "Enclosure 0, Fan Module 3",
-						"status-ses": "OK",
-						"status-ses-numeric": 1,
-						"extended-status": "00000010",
-						"status": "Up",
-						"status-numeric": 0,
-						"speed": 13860,
-						"position": "N/A",
-						"position-numeric": 6,
-						"serial-number": "N/A",
-						"part-number": "N/A",
-						"fw-revision": "",
-						"hw-revision": "",
-						"locator-led": "Off",
-						"locator-led-numeric": 0,
-						"health": "OK",
-						"health-numeric": 0,
-						"health-reason": "",
-						"health-recommendation": ""
-					}
-				]
-			},
-			{
-				"object-name": "fan-modules",
-				"meta": "/meta/fan-modules",
-				"durable-id": "fan_module_0.4",
-				"enclosure-id": 0,
-				"dom-id": 4,
-				"name": "Fan Module 4",
-				"location": "Enclosure 0 - Right",
-				"status": "OK",
-				"status-numeric": 1,
-				"extended-status": "00000000",
-				"position": "Indexed",
-				"position-numeric": 5,
-				"health": "Degraded",
-				"health-numeric": 1,
-				"health-reason": "A subcomponent of this component is unhealthy.",
-				"health-recommendation": "- See the health information for the unhealthy subcomponents of this component.",
-				"fan": [
-					{
-						"object-name": "fan-details",
-						"meta": "/meta/fan",
-						"durable-id": "fan_0.fm4.0",
-						"name": "Fan 0",
-						"location": "Enclosure 0, Fan Module 4",
-						"status-ses": "OK",
-						"status-ses-numeric": 1,
-						"extended-status": "00000010",
-						"status": "Up",
-						"status-numeric": 0,
-						"speed": 6000,
-						"position": "N/A",
-						"position-numeric": 6,
-						"serial-number": "N/A",
-						"part-number": "N/A",
-						"fw-revision": "",
-						"hw-revision": "",
-						"locator-led": "Off",
-						"locator-led-numeric": 0,
-						"health": "OK",
-						"health-numeric": 0,
-						"health-reason": "",
-						"health-recommendation": ""
-					},
-					{
-						"object-name": "fan-details",
-						"meta": "/meta/fan",
-						"durable-id": "fan_0.fm4.1",
-						"name": "Fan 1",
-						"location": "Enclosure 0, Fan Module 4",
-						"status-ses": "OK",
-						"status-ses-numeric": 1,
-						"extended-status": "00000010",
-						"status": "Up",
-						"status-numeric": 0,
-						"speed": 6480,
-						"position": "N/A",
-						"position-numeric": 6,
-						"serial-number": "N/A",
-						"part-number": "N/A",
-						"fw-revision": "",
-						"hw-revision": "",
-						"locator-led": "Off",
-						"locator-led-numeric": 0,
-						"health": "OK",
-						"health-numeric": 0,
-						"health-reason": "",
-						"health-recommendation": ""
-					}
-				]
-			}
-		],
-		"status": [
-			{
-				"object-name": "status",
-				"meta": "/meta/status",
-				"response-type": "Success",
-				"response-type-numeric": 0,
-				"response": "Command completed successfully. (2019-08-02 11:05:55)",
-				"return-code": 0,
-				"component-id": "",
-				"time-stamp": "2019-08-02 11:05:55",
-				"time-stamp-numeric": 1564743955
-			}
-		]
-	}}
-    """
-
-PSU_RESPONSE = """
-    {
-    "status_code": 200,
-    "api-response": {
-    "power-supplies": [
-        {
-            "object-name": "power-supplies",
-            "durable-id": "psu_0.0",
-            "enclosure-id": 0,
-            "dom-id": 0,
-            "serial-number": "",
-            "part-number": "",
-            "description": "",
-            "name": "PSU 0, Left",
-            "fw-revision": "0229",
-            "revision": "",
-            "model": "",
-            "vendor": "",
-            "location": "Enclosure 0 - Left",
-            "position": "Left",
-            "position-numeric": 0,
-            "dash-level": "",
-            "fru-shortname": "",
-            "mfg-date": "N/A",
-            "mfg-date-numeric": 0,
-            "mfg-location": "",
-            "mfg-vendor-id": "",
-            "configuration-serialnumber": "",
-            "dc12v": 0,
-            "dc5v": 0,
-            "dc33v": 0,
-            "dc12i": 0,
-            "dc5i": 0,
-            "dctemp": 0,
-            "health": "Degraded",
-            "health-numeric": 1,
-            "health-reason": "Unable to obtain FRU ID data for this FRU.",
-            "health-recommendation": "- The FRU ID content was not read. This could be temporary or it may not programmed correctly.     - If this persists, contact technical support.",
-            "status": "Up",
-            "status-numeric": 0
-        },
-        {
-            "object-name": "power-supplies",
-            "durable-id": "psu_0.1",
-            "enclosure-id": 0,
-            "dom-id": 1,
-            "serial-number": "",
-            "part-number": "",
-            "description": "",
-            "name": "PSU 1, Right",
-            "fw-revision": "0229",
-            "revision": "",
-            "model": "",
-            "vendor": "",
-            "location": "Enclosure 0 - Right",
-            "position": "Right",
-            "position-numeric": 1,
-            "dash-level": "",
-            "fru-shortname": "",
-            "mfg-date": "N/A",
-            "mfg-date-numeric": 0,
-            "mfg-location": "",
-            "mfg-vendor-id": "",
-            "configuration-serialnumber": "",
-            "dc12v": 0,
-            "dc5v": 0,
-            "dc33v": 0,
-            "dc12i": 0,
-            "dc5i": 0,
-            "dctemp": 0,
-            "health": "OK",
-            "health-numeric": 1,
-            "health-reason": "Unable to obtain FRU ID data for this FRU.",
-            "health-recommendation": "- The FRU ID content was not read. This could be temporary or it may not programmed correctly.     - If this persists, contact technical support.",
-            "status": "Up",
-            "status-numeric": 0
-        }
-    ],
-    "status": [
-        {
-            "object-name": "status",
-            "response-type": "Success",
-            "response-type-numeric": 0,
-            "response": "Command completed successfully. (2019-07-04 04:23:04)",
-            "return-code": 0,
-            "component-id": "",
-            "time-stamp": "2019-07-04 04:23:04",
-            "time-stamp-numeric": 1562214184
-        }
-    ]
-    }}
-    """
