+++ unittests/files/opt/seagate/sspl/setup/resource_map/lr2/encl_api_response.py
@@ -919,3 +919,468 @@ SAS_PORTS_RESPONSE = [{
         "health-reason": "",
         "health-recommendation": ""
     }]
+
+FAN_MODULE_RESPONSE = """
+    {
+	"status_code": 200,
+	"api-response": {
+		"fan-modules": [
+			{
+				"object-name": "fan-modules",
+				"meta": "/meta/fan-modules",
+				"durable-id": "fan_module_0.0",
+				"enclosure-id": 0,
+				"dom-id": 0,
+				"name": "Fan Module 0",
+				"location": "Enclosure 0 - Left",
+				"status": "OK",
+				"status-numeric": 1,
+				"extended-status": "00000000",
+				"position": "Indexed",
+				"position-numeric": 5,
+				"health": "OK",
+				"health-numeric": 0,
+				"health-reason": "",
+				"health-recommendation": "",
+				"fan": [
+					{
+						"object-name": "fan-details",
+						"meta": "/meta/fan",
+						"durable-id": "fan_0.fm0.0",
+						"name": "Fan 0",
+						"location": "Enclosure 0, Fan Module 0",
+						"status-ses": "OK",
+						"status-ses-numeric": 1,
+						"extended-status": "00000010",
+						"status": "Up",
+						"status-numeric": 0,
+						"speed": 13800,
+						"position": "N/A",
+						"position-numeric": 6,
+						"serial-number": "N/A",
+						"part-number": "N/A",
+						"fw-revision": "",
+						"hw-revision": "",
+						"locator-led": "Off",
+						"locator-led-numeric": 0,
+						"health": "OK",
+						"health-numeric": 0,
+						"health-reason": "",
+						"health-recommendation": ""
+					},
+					{
+						"object-name": "fan-details",
+						"meta": "/meta/fan",
+						"durable-id": "fan_0.fm0.1",
+						"name": "Fan 1",
+						"location": "Enclosure 0, Fan Module 0",
+						"status-ses": "OK",
+						"status-ses-numeric": 1,
+						"extended-status": "00000010",
+						"status": "Up",
+						"status-numeric": 0,
+						"speed": 13860,
+						"position": "N/A",
+						"position-numeric": 6,
+						"serial-number": "N/A",
+						"part-number": "N/A",
+						"fw-revision": "",
+						"hw-revision": "",
+						"locator-led": "Off",
+						"locator-led-numeric": 0,
+						"health": "OK",
+						"health-numeric": 0,
+						"health-reason": "",
+						"health-recommendation": ""
+					}
+				]
+			},
+			{
+				"object-name": "fan-modules",
+				"meta": "/meta/fan-modules",
+				"durable-id": "fan_module_0.1",
+				"enclosure-id": 0,
+				"dom-id": 1,
+				"name": "Fan Module 1",
+				"location": "Enclosure 0 - Left-Middle",
+				"status": "OK",
+				"status-numeric": 1,
+				"extended-status": "00000000",
+				"position": "Indexed",
+				"position-numeric": 5,
+				"health": "OK",
+				"health-numeric": 0,
+				"health-reason": "",
+				"health-recommendation": "",
+				"fan": [
+					{
+						"object-name": "fan-details",
+						"meta": "/meta/fan",
+						"durable-id": "fan_0.fm1.0",
+						"name": "Fan 0",
+						"location": "Enclosure 0, Fan Module 1",
+						"status-ses": "OK",
+						"status-ses-numeric": 1,
+						"extended-status": "00000010",
+						"status": "Up",
+						"status-numeric": 0,
+						"speed": 13860,
+						"position": "N/A",
+						"position-numeric": 6,
+						"serial-number": "N/A",
+						"part-number": "N/A",
+						"fw-revision": "",
+						"hw-revision": "",
+						"locator-led": "Off",
+						"locator-led-numeric": 0,
+						"health": "OK",
+						"health-numeric": 0,
+						"health-reason": "",
+						"health-recommendation": ""
+					},
+					{
+						"object-name": "fan-details",
+						"meta": "/meta/fan",
+						"durable-id": "fan_0.fm1.1",
+						"name": "Fan 1",
+						"location": "Enclosure 0, Fan Module 1",
+						"status-ses": "OK",
+						"status-ses-numeric": 1,
+						"extended-status": "00000010",
+						"status": "Up",
+						"status-numeric": 0,
+						"speed": 13500,
+						"position": "N/A",
+						"position-numeric": 6,
+						"serial-number": "N/A",
+						"part-number": "N/A",
+						"fw-revision": "",
+						"hw-revision": "",
+						"locator-led": "Off",
+						"locator-led-numeric": 0,
+						"health": "OK",
+						"health-numeric": 0,
+						"health-reason": "",
+						"health-recommendation": ""
+					}
+				]
+			},
+			{
+				"object-name": "fan-modules",
+				"meta": "/meta/fan-modules",
+				"durable-id": "fan_module_0.2",
+				"enclosure-id": 0,
+				"dom-id": 2,
+				"name": "Fan Module 2",
+				"location": "Enclosure 0 - Middle",
+				"status": "OK",
+				"status-numeric": 1,
+				"extended-status": "00000000",
+				"position": "Indexed",
+				"position-numeric": 5,
+				"health": "OK",
+				"health-numeric": 0,
+				"health-reason": "",
+				"health-recommendation": "",
+				"fan": [
+					{
+						"object-name": "fan-details",
+						"meta": "/meta/fan",
+						"durable-id": "fan_0.fm2.0",
+						"name": "Fan 0",
+						"location": "Enclosure 0, Fan Module 2",
+						"status-ses": "OK",
+						"status-ses-numeric": 1,
+						"extended-status": "00000010",
+						"status": "Up",
+						"status-numeric": 0,
+						"speed": 13860,
+						"position": "N/A",
+						"position-numeric": 6,
+						"serial-number": "N/A",
+						"part-number": "N/A",
+						"fw-revision": "",
+						"hw-revision": "",
+						"locator-led": "Off",
+						"locator-led-numeric": 0,
+						"health": "OK",
+						"health-numeric": 0,
+						"health-reason": "",
+						"health-recommendation": ""
+					},
+					{
+						"object-name": "fan-details",
+						"meta": "/meta/fan",
+						"durable-id": "fan_0.fm2.1",
+						"name": "Fan 1",
+						"location": "Enclosure 0, Fan Module 2",
+						"status-ses": "OK",
+						"status-ses-numeric": 1,
+						"extended-status": "00000010",
+						"status": "Up",
+						"status-numeric": 0,
+						"speed": 13860,
+						"position": "N/A",
+						"position-numeric": 6,
+						"serial-number": "N/A",
+						"part-number": "N/A",
+						"fw-revision": "",
+						"hw-revision": "",
+						"locator-led": "Off",
+						"locator-led-numeric": 0,
+						"health": "OK",
+						"health-numeric": 0,
+						"health-reason": "",
+						"health-recommendation": ""
+					}
+				]
+			},
+			{
+				"object-name": "fan-modules",
+				"meta": "/meta/fan-modules",
+				"durable-id": "fan_module_0.3",
+				"enclosure-id": 0,
+				"dom-id": 3,
+				"name": "Fan Module 3",
+				"location": "Enclosure 0 - Right-Middle",
+				"status": "OK",
+				"status-numeric": 1,
+				"extended-status": "00000000",
+				"position": "Indexed",
+				"position-numeric": 5,
+				"health": "OK",
+				"health-numeric": 0,
+				"health-reason": "",
+				"health-recommendation": "",
+				"fan": [
+					{
+						"object-name": "fan-details",
+						"meta": "/meta/fan",
+						"durable-id": "fan_0.fm3.0",
+						"name": "Fan 0",
+						"location": "Enclosure 0, Fan Module 3",
+						"status-ses": "OK",
+						"status-ses-numeric": 1,
+						"extended-status": "00000010",
+						"status": "Up",
+						"status-numeric": 0,
+						"speed": 13860,
+						"position": "N/A",
+						"position-numeric": 6,
+						"serial-number": "N/A",
+						"part-number": "N/A",
+						"fw-revision": "",
+						"hw-revision": "",
+						"locator-led": "Off",
+						"locator-led-numeric": 0,
+						"health": "OK",
+						"health-numeric": 0,
+						"health-reason": "",
+						"health-recommendation": ""
+					},
+					{
+						"object-name": "fan-details",
+						"meta": "/meta/fan",
+						"durable-id": "fan_0.fm3.1",
+						"name": "Fan 1",
+						"location": "Enclosure 0, Fan Module 3",
+						"status-ses": "OK",
+						"status-ses-numeric": 1,
+						"extended-status": "00000010",
+						"status": "Up",
+						"status-numeric": 0,
+						"speed": 13860,
+						"position": "N/A",
+						"position-numeric": 6,
+						"serial-number": "N/A",
+						"part-number": "N/A",
+						"fw-revision": "",
+						"hw-revision": "",
+						"locator-led": "Off",
+						"locator-led-numeric": 0,
+						"health": "OK",
+						"health-numeric": 0,
+						"health-reason": "",
+						"health-recommendation": ""
+					}
+				]
+			},
+			{
+				"object-name": "fan-modules",
+				"meta": "/meta/fan-modules",
+				"durable-id": "fan_module_0.4",
+				"enclosure-id": 0,
+				"dom-id": 4,
+				"name": "Fan Module 4",
+				"location": "Enclosure 0 - Right",
+				"status": "OK",
+				"status-numeric": 1,
+				"extended-status": "00000000",
+				"position": "Indexed",
+				"position-numeric": 5,
+				"health": "Degraded",
+				"health-numeric": 1,
+				"health-reason": "A subcomponent of this component is unhealthy.",
+				"health-recommendation": "- See the health information for the unhealthy subcomponents of this component.",
+				"fan": [
+					{
+						"object-name": "fan-details",
+						"meta": "/meta/fan",
+						"durable-id": "fan_0.fm4.0",
+						"name": "Fan 0",
+						"location": "Enclosure 0, Fan Module 4",
+						"status-ses": "OK",
+						"status-ses-numeric": 1,
+						"extended-status": "00000010",
+						"status": "Up",
+						"status-numeric": 0,
+						"speed": 6000,
+						"position": "N/A",
+						"position-numeric": 6,
+						"serial-number": "N/A",
+						"part-number": "N/A",
+						"fw-revision": "",
+						"hw-revision": "",
+						"locator-led": "Off",
+						"locator-led-numeric": 0,
+						"health": "OK",
+						"health-numeric": 0,
+						"health-reason": "",
+						"health-recommendation": ""
+					},
+					{
+						"object-name": "fan-details",
+						"meta": "/meta/fan",
+						"durable-id": "fan_0.fm4.1",
+						"name": "Fan 1",
+						"location": "Enclosure 0, Fan Module 4",
+						"status-ses": "OK",
+						"status-ses-numeric": 1,
+						"extended-status": "00000010",
+						"status": "Up",
+						"status-numeric": 0,
+						"speed": 6480,
+						"position": "N/A",
+						"position-numeric": 6,
+						"serial-number": "N/A",
+						"part-number": "N/A",
+						"fw-revision": "",
+						"hw-revision": "",
+						"locator-led": "Off",
+						"locator-led-numeric": 0,
+						"health": "OK",
+						"health-numeric": 0,
+						"health-reason": "",
+						"health-recommendation": ""
+					}
+				]
+			}
+		],
+		"status": [
+			{
+				"object-name": "status",
+				"meta": "/meta/status",
+				"response-type": "Success",
+				"response-type-numeric": 0,
+				"response": "Command completed successfully. (2019-08-02 11:05:55)",
+				"return-code": 0,
+				"component-id": "",
+				"time-stamp": "2019-08-02 11:05:55",
+				"time-stamp-numeric": 1564743955
+			}
+		]
+	}}
+    """
+
+PSU_RESPONSE = """
+    {
+    "status_code": 200,
+    "api-response": {
+    "power-supplies": [
+        {
+            "object-name": "power-supplies",
+            "durable-id": "psu_0.0",
+            "enclosure-id": 0,
+            "dom-id": 0,
+            "serial-number": "",
+            "part-number": "",
+            "description": "",
+            "name": "PSU 0, Left",
+            "fw-revision": "0229",
+            "revision": "",
+            "model": "",
+            "vendor": "",
+            "location": "Enclosure 0 - Left",
+            "position": "Left",
+            "position-numeric": 0,
+            "dash-level": "",
+            "fru-shortname": "",
+            "mfg-date": "N/A",
+            "mfg-date-numeric": 0,
+            "mfg-location": "",
+            "mfg-vendor-id": "",
+            "configuration-serialnumber": "",
+            "dc12v": 0,
+            "dc5v": 0,
+            "dc33v": 0,
+            "dc12i": 0,
+            "dc5i": 0,
+            "dctemp": 0,
+            "health": "Degraded",
+            "health-numeric": 1,
+            "health-reason": "Unable to obtain FRU ID data for this FRU.",
+            "health-recommendation": "- The FRU ID content was not read. This could be temporary or it may not programmed correctly.     - If this persists, contact technical support.",
+            "status": "Up",
+            "status-numeric": 0
+        },
+        {
+            "object-name": "power-supplies",
+            "durable-id": "psu_0.1",
+            "enclosure-id": 0,
+            "dom-id": 1,
+            "serial-number": "",
+            "part-number": "",
+            "description": "",
+            "name": "PSU 1, Right",
+            "fw-revision": "0229",
+            "revision": "",
+            "model": "",
+            "vendor": "",
+            "location": "Enclosure 0 - Right",
+            "position": "Right",
+            "position-numeric": 1,
+            "dash-level": "",
+            "fru-shortname": "",
+            "mfg-date": "N/A",
+            "mfg-date-numeric": 0,
+            "mfg-location": "",
+            "mfg-vendor-id": "",
+            "configuration-serialnumber": "",
+            "dc12v": 0,
+            "dc5v": 0,
+            "dc33v": 0,
+            "dc12i": 0,
+            "dc5i": 0,
+            "dctemp": 0,
+            "health": "OK",
+            "health-numeric": 1,
+            "health-reason": "Unable to obtain FRU ID data for this FRU.",
+            "health-recommendation": "- The FRU ID content was not read. This could be temporary or it may not programmed correctly.     - If this persists, contact technical support.",
+            "status": "Up",
+            "status-numeric": 0
+        }
+    ],
+    "status": [
+        {
+            "object-name": "status",
+            "response-type": "Success",
+            "response-type-numeric": 0,
+            "response": "Command completed successfully. (2019-07-04 04:23:04)",
+            "return-code": 0,
+            "component-id": "",
+            "time-stamp": "2019-07-04 04:23:04",
+            "time-stamp-numeric": 1562214184
+        }
+    ]
+    }}
+    """
